+++ src/Modules/NewsalertSubscribeModule.php
@@ -77,8 +77,6 @@ class NewsalertSubscribeModule extends \Module
         $this->formHybridDataContainer = static::TABLE;
         $this->formHybridAsync = '1';
 
-//        'formHybridOptInSuccessMessage,formHybridOptInNotification,formHybridOptInConfirmedProperty'
-
         if ($this->newsalertOptIn) {
             $this->formHybridAddOptIn = 1;
             $this->formHybridOptInConfirmedProperty = 'confirmed';